% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_optimal_density_estimate.R
\name{cv_optimal_density_estimate}
\alias{cv_optimal_density_estimate}
\title{Cross Validation to Choose the Best Penalty Parameter}
\usage{
cv_optimal_density_estimate(
  data,
  domain,
  penalty_param_candidates,
  fold_number = 5
)
}
\arguments{
\item{data}{A numeric vector whose log-concave density function is to be estimated;
missing values are automatically removed.}

\item{domain}{A numeric vector of length 2 specifying the left and right
endpoints of the bounded domain;
its components cannot be \code{NA}, \code{NULL}, \code{-Inf}, or \code{Inf}.}

\item{penalty_param_candidates}{A numeric vector of the penalty parameter candidates;
each element must be non-negative.}

\item{fold_number}{An integer to indicate the number of folds for cross validation.
Default is \code{5}.}
}
\value{
A object with class "LogConcaveDESM" with the penalty parameter being the optimal choice.
}
\description{
Given a numeric vector of penalty parameter candidates, use the cross validation to
choose the best one so that the resulting density estimate not only fits the training data well
but can also generalize to unseen data.
}
\examples{
set.seed(1119)
N <- 100
data <- rnorm(N)
domain <- c(-5, 5)
lambda_cand <- exp(c(-Inf, seq(-3, 1, by = 0.5)))

opt_result <- cv_optimal_density_estimate(data, domain, lambda_cand)

}
\seealso{
\code{\link{lcd_scorematching}}
}
