% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_metrics.R
\name{density_metrics}
\alias{density_metrics}
\alias{kl_div}
\alias{hyvarinen_div}
\alias{L1_dist}
\alias{hellinger_dist}
\title{Computation of Various Distances between Two Probability Distributions}
\usage{
kl_div(
  true_density,
  density_estimate,
  minus_const = 0,
  mc_batch_size = 100,
  mc_rel_tol_param = 0.01,
  print_error = FALSE
)

hyvarinen_div(
  true_density,
  density_estimate,
  mc_batch_size = 100,
  mc_rel_tol_param = 0.01,
  print_error = FALSE
)

L1_dist(
  true_density,
  density_estimate,
  minus_const = 0,
  mc_batch_size = 100,
  mc_rel_tol_param = 0.01,
  print_error = FALSE
)

hellinger_dist(
  true_density,
  density_estimate,
  minus_const = 0,
  mc_batch_size = 100,
  mc_rel_tol_param = 0.01,
  print_error = FALSE
)
}
\arguments{
\item{true_density}{An \code{R} object of class "truncated_normal", "truncated_gamma",
"truncated_lognormal" or "beta_dist", returned from the \code{truncated_normal},
\code{truncated_gamma}, \code{truncated_lognormal}, or \code{beta_dist}, respectively.}

\item{density_estimate}{An object of class "LogConcaveDESM",
usually the output of \code{\link{lcd_scorematching}} or
\code{\link{cv_optimal_density_estimate}}.}

\item{minus_const}{A numeric to be subtracted in the exponent to
ensure the finite-ness of the integration result. Default is \code{0}.}

\item{mc_batch_size}{A numeric to specify the batch size of each Monte Carlo draw of random samples.
Default is \code{100}.}

\item{mc_rel_tol_param}{A numeric to specify the criterion of terminating the batch Monte Carlo algorithm.
Default is \code{1e-2}.}

\item{print_error}{A logical value to specify whether to print the error after each Monte Carlo draw.
Default is \code{FALSE}.}
}
\value{
The approximate value of the probability distribution distance.
}
\description{
A collection of functions to compute various distances between two probability distributions,
which are used to assess the quality of the density estimate.
}
