% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logden_deriv1.R
\name{evaluate_logden_deriv1}
\alias{evaluate_logden_deriv1}
\alias{plot_logden_deriv1}
\title{Computation and Plotting the First Derivative of the Logarithm of the Log-concave Score Matching Density Estimate}
\usage{
evaluate_logden_deriv1(scorematching_logconcave, newx)

plot_logden_deriv1(
  scorematching_logconcave,
  plot_domain,
  plot_points_cnt = 100
)
}
\arguments{
\item{scorematching_logconcave}{An object of class "LogConcaveDESM",
usually the output of \code{\link{lcd_scorematching}} or \code{\link{cv_optimal_density_estimate}}.}

\item{newx}{A numeric vector of real numbers at which the first derivative of the logarithm of
the log-concave score matching density estimate should be evaluated.}

\item{plot_domain}{A numeric vector to indicate the domain of the plot.}

\item{plot_points_cnt}{A numeric to indicate the number of points for evaluating and plotting.
Default is \code{100}.}
}
\value{
A data frame with the first column being the sorted newx and
the second column being the corresponding first derivative values of the logarithm of
the log-concave score matching density estimate.

A ggplot2 plot of the first derivative of the logarithm of
the log-concave score matching density estimate over the specified plot domain.
}
\description{
Based on a "LogConcaveDESM" object, evaluates and plots the first derivative of the logarithm of
the log-concave score matching density estimate.
}
\examples{
set.seed(1119)
N <- 100
data <- rnorm(N)
domain <- c(-5, 5)
result <- lcd_scorematching(data, domain, penalty_param = 1e-10)
# evaluationn
evaluate_logden_deriv1(scorematching_logconcave = result,
newx = seq(result$domain[1], result$domain[2], 0.01))

# plot
plot_logden_deriv1(scorematching_logconcave = result,
plot_domain = result$domain, plot_points_cnt = 500)

}
