% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_scorematching_loss.R
\name{evaluate_scorematching_loss}
\alias{evaluate_scorematching_loss}
\title{Evaluation of the Score Matching Loss Function}
\usage{
evaluate_scorematching_loss(scorematching_logconcave, new_data)
}
\arguments{
\item{scorematching_logconcave}{An object of class "LogConcaveDESM",
usually the output of \code{\link{lcd_scorematching}} or \code{\link{cv_optimal_density_estimate}}.}

\item{new_data}{A numeric vector of real numbers at which the score matching loss function
should be evaluated.}
}
\value{
A numeric of the score matching loss function evaluated at the new_data.
}
\description{
Evaluates the score matching loss function at \code{new_data}.
}
\examples{
set.seed(1119)
N <- 100
data <- rnorm(N)
domain <- c(-5, 5)
result <- lcd_scorematching(data, domain, penalty_param = 1e-10)

evaluate_scorematching_loss(scorematching_logconcave = result, new_data = result$data)

evaluate_scorematching_loss(scorematching_logconcave = result,
new_data = seq(-5, 5, by = 0.01))

}
