% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_expcubic.R
\name{integrate_expcubic}
\alias{integrate_expcubic}
\title{Integration of an Exponential-Cubic Function}
\usage{
integrate_expcubic(a, b, c, d, lower, upper, minus_const = 0)
}
\arguments{
\item{a, b, c, d}{The coefficients of \eqn{x^3}, \eqn{x^2}, \eqn{x}, and the constant term, respectively.}

\item{lower, upper}{Numeric to specify the limits of integration; must be finite.}

\item{minus_const}{A numeric to be subtracted in the exponent to
ensure the finite-ness of the integration result. Default is \code{0}.}
}
\value{
A numeric of the integral of \eqn{\exp(a*x^3+b*x^2+c*x+d)} from \code{lower} to \code{upper}.
}
\description{
Integrates a function of the form \eqn{\exp(a*x^3+b*x^2+c*x+d)} over a bounded interval.
}
\examples{
integrate_expcubic(a = 1, b = -2, c = 3, d = -2, lower = -2, upper = 2)

}
