% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcd_scorematching.R
\name{lcd}
\alias{lcd}
\alias{lcd_scorematching}
\alias{lcd_scorematching_bounded}
\alias{lcd_scorematching_R}
\alias{lcd_scorematching_ninfb}
\alias{lcd_scorematching_ainf}
\title{Computation of Second Derivative of Score Matching Log-concave Density Estimate}
\usage{
lcd_scorematching(data, domain, penalty_param = 0.1, verbose = FALSE)

lcd_scorematching_bounded(data, domain, penalty_param = 0.1, verbose = FALSE)

lcd_scorematching_R(data, domain, penalty_param = 0.1, verbose = FALSE)

lcd_scorematching_ninfb(data, domain, penalty_param = 0.1, verbose = FALSE)

lcd_scorematching_ainf(data, domain, penalty_param = 0.1, verbose = FALSE)
}
\arguments{
\item{data}{A numeric vector whose log-concave density function is to be estimated;
missing values are automatically removed.}

\item{domain}{A numeric vector of length 2 specifying the left and right
endpoints of the domain. The \code{domain} can be a bounded interval, the entire real line,
a left-bounded and right-unbounded interval, or a left-unbounded and right-bounded interval.
Its components cannot be \code{NaN}.}

\item{penalty_param}{Penalty parameter for computing the density estimate; must be non-negative.
Default is \code{1e-1}.}

\item{verbose}{An optional logical value to indicate whether to print detailed CVXR
solver output. Default is \code{FALSE}.}
}
\value{
An object of class "LogConcaveDESM" whose underlying structure is
a list containing the following elements
\describe{
    \item{call}{call: the call which produced the result.}
    \item{data}{numeric: the original data whose log-concave density function is to be estimated.}
    \item{sorted_unique_data}{numeric: the sorted unique data.}
    \item{data_weights}{numeric: weight vector of the sorted_unique_data.}
    \item{domain}{numeric: the bounded domain over the density function is assumed and estimated.}
    \item{opt_theta}{numeric: the optimal second derivative of the logarithm of the (penalized)
    log-concave score matching density estimate at the sorted unique data and two boundary points.}
    \item{matrix_A}{numeric: matrix A used to compute the log-concave score matching density estimate.}
    \item{matrix_B}{numeric: matrix B used to compute the log-concave score matching density estimate.}
    \item{penalty_param}{numeric: the penalty parameter used to compute the log-concave
    score matching density estimate.}
}
}
\description{
Computes the second derivative of the logarithm of the (penalized) log-concave score matching
density estimate based on i.i.d samples, assuming it is a continuous piecewise linear function
with knots at the samples.
The nature of the underlying density estimation problem is a constrained quadratic optimization problem,
which is solved by \code{CVXR} package.
The output is an object of class "\code{LogConcaveDESM}".
}
\details{
The functions \code{lcd_scorematching_bounded}, \code{lcd_scorematching_R}, \code{lcd_scorematching_ninfb} and
\code{lcd_scorematching_ainf} compute the second derivative of the logarithm of the (penalized) log-concave score matching density estimate
at \code{data} when the underlying \code{domain} is a bounded interval, the entire real line,
an interval of the form \eqn{-\infty, b} for some \eqn{b < \infty}, and
an interval of the form \eqn{a, \infty} for some \eqn{a > -\infty}, respectively.
The function \code{lcd_scorematching} encompasses all four cases.
}
\examples{
set.seed(1119)
N <- 100
data <- rnorm(N)
domain <- c(-5, 5)
# no penalty term
result <- lcd_scorematching(data, domain, penalty_param = 0)

# with penalty term
result <- lcd_scorematching(data, domain, penalty_param = 1e-10)

}
