% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogConcaveDESM.R
\docType{package}
\name{LogConcaveDESM-package}
\alias{LogConcaveDESM-package}
\title{LogConcaveDESM: a package for log-concave density estimation using the score matching loss function}
\description{
\code{LogConcaveDESM} is an \code{R} package to compute and
visualize the log-concave score matching density estimate over a bounded domain.
It also contains functions to plot the first and second derivatives of the log-density estimate.
Furthermore, functions to compute various distances between two probability distributions
are provided to assess the quality of the density estimate.
}
\details{
LogConcaveDESM contains functions to compute and visualize the log-concave
score matching density estimate.

\code{\link{lcd_scorematching}} computes the second derivative of
the (penalized) log-concave score matching density estimate based on i.i.d samples.
The output is an object of class "\code{LogConcaveDESM}" which is used as an input to
various auxiliary functions.

\code{\link{evaluate_logdensity_deriv2}} and \code{\link{plot_logdensity_deriv2}}
evaluates and plots the second derivative of the logarithm of the log-concave score matching
density estimate, respectively.

\code{\link{evaluate_logdensity_deriv1}} evaluates and \code{\link{plot_logdensity_deriv1}}
plots the first derivative of the logarithm of the log-concave score matching
density estimate, respectively.

\code{\link{evaluate_logdensity}} evaluates and \code{\link{plot_logdensity}} plots
the logarithm of the log-concave score matching density estimate up to a normalizing constant, respectively.

\code{\link{evaluate_density}} evaluates and \code{\link{plot_density}} plots the log-concave
score matching density estimate, respectively.

\code{\link{cv_optimal_density_estimate}} chooses the optimal penalty parameter.

\code{\link{plot_mle_scorematching}} plots the log-concave maximum likelihood and score matching
density estimates together.

\code{\link{kl_div}}, \code{\link{hyvarinen_div}}, \code{\link{L1_dist}} and
 \code{\link{hellinger_dist}} compute the Kullback-Leibler divergence, Hyvarinen divergence,
 L1 distance, and Hellinger distance between the true density function and
 the density estimate, respectively.
}
\seealso{
\href{https://cran.r-project.org/web/packages/logcondens/index.html}{logcondens},
\href{https://cran.r-project.org/web/packages/LogConcDEAD/index.html}{LogConcDEAD}.
}
