% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizing_const.R
\name{normalizing_const}
\alias{normalizing_const}
\title{Computation the Normalizing Constant of a Log-concave Score Matching Density Estimate}
\usage{
normalizing_const(scorematching_logconcave, minus_const = 0)
}
\arguments{
\item{scorematching_logconcave}{An object of class "LogConcaveDESM",
usually the output of \code{\link{lcd_scorematching}} or \code{\link{cv_optimal_density_estimate}}.}

\item{minus_const}{A numeric to be subtracted in the exponent to
ensure the finite-ness of the integration result. Default is \code{0}.}
}
\value{
The normalizing constant of a log-concave score matching density estimate.
}
\description{
Computes the normalizing constant of a log-concave score matching density estimate.
}
\examples{
set.seed(1119)
N <- 100
data <- rnorm(N)
domain <- c(-5, 5)
# no penalty term
result <- lcd_scorematching(data, domain, penalty_param = 0)
normalizing_const(result, minus_const = 500)

}
