% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizing_const.R
\name{normalizing_const}
\alias{normalizing_const}
\alias{normalizing_const_bounded}
\alias{normalizing_const_R}
\alias{normalizing_const_ninfb}
\alias{normalizing_const_ainf}
\title{Computation the Normalizing Constant}
\usage{
normalizing_const(scorematching_logconcave, minus_const = 0)

normalizing_const_bounded(scorematching_logconcave, minus_const = 0)

normalizing_const_R(scorematching_logconcave, minus_const = 0)

normalizing_const_ninfb(scorematching_logconcave, minus_const = 0)

normalizing_const_ainf(scorematching_logconcave, minus_const = 0)
}
\arguments{
\item{scorematching_logconcave}{An object of class "LogConcaveDESM",
usually the output of \code{\link{lcd_scorematching}} or \code{\link{cv_optimal_density_estimate}}.}

\item{minus_const}{A numeric to be subtracted in the exponent to
ensure the finite-ness of the integration result. Default is \code{0}.}
}
\value{
The normalizing constant of a (penalized) un-normalized log-concave score matching density estimate.
}
\description{
Computes the normalizing constant of a (penalized) un-normalized log-concave score matching density estimate.
}
\details{
The functions \code{normalizing_const_bounded}, \code{normalizing_const_R}, \code{normalizing_const_ninfb},
and \code{normalizing_const_ainf} computes the normalizing constant of a (penalized) un-normalized
log-concave score matching density estimate when the underlying \code{domain} is a bounded interval,
the entire real line, an interval of the form \eqn{(-\infty, b)}
for some \eqn{b < \infty}, and an interval of the form \eqn{(a, \infty)} for some \eqn{a > -\infty}, respectively.
The function \code{normalizing_const} encompasses all four cases.
}
\examples{
set.seed(1119)
N <- 100
data <- rnorm(N)
domain <- c(-5, 5)
# no penalty term
result <- lcd_scorematching(data, domain, penalty_param = 0)
normalizing_const(result, minus_const = 500)

}
